Fix: Merge traits
